reducerと個々のdispatcherを別ファイルに切り分ける
